requirements.txt
langchain
langchain-openai
langchain-community
langchain-chroma
pypdf
chromadb
python-dotenv
tiktoken
tavily-python
chainlit